# yuihub_api/openapi.yml
openapi: 3.1.1
info:
  title: YuiHub API (YuiFlow Ph2b)
  description: |
    YuiHub provides GPTs⇄Copilot bridging through YuiFlow specification compliance.
    Enables saving, searching, and context packet generation for AI conversations 
    across multiple AI clients with complete YuiFlow schema support.
  version: 0.2.0
  contact:
    name: YuiHub Support
  license:
    name: MIT
  termsOfService: https://poc-yuihub.vemi.jp/privacy
  x-privacy-policy: https://poc-yuihub.vemi.jp/privacy
servers:
  - url: https://poc-yuihub.vemi.jp
    description: Production server (Fixed Tunnel)
  - url: http://localhost:3000
    description: Local development server

# Global security requirement
security:
  - ApiKeyAuth: []
paths:
...
  /save:
    post:
      summary: Save message fragment using YuiFlow InputMessage format
      description: |
        Saves a message fragment using YuiFlow specification. Accepts InputMessage 
        format and converts to Fragment for storage. Supports GPTs⇄Copilot bridging.
      operationId: saveMessage
      x-openai-isConsequential: false
      tags:
        - YuiFlow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputMessage'
            examples:
              gpts_message:
                summary: GPTs message example
                value:
                  source: "gpts"
                  thread: "th-01K5WHS123EXAMPLE456789ABC"
                  author: "ChatGPT"
                  text: "新機能の設計について議論しました。ユーザビリティを重視した実装を採用します。"
                  tags: ["design", "feature", "decision"]
      responses:
        '200':
          description: Successfully saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '400':
          description: Invalid InputMessage format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /search:
    get:
      summary: Enhanced search with YuiFlow filters
      description: |
        Full-text search with YuiFlow-compliant filtering by tag, thread, 
        source, and author. Returns ranked results with metadata.
      operationId: searchMessages
      tags:
        - YuiFlow
        - Search
      parameters:
        - name: q
          in: query
          description: Text search query (optional if using filters)
          schema:
            type: string
            example: "設計 決定"
        - name: tag
          in: query
          description: Filter by specific tag
          schema:
            type: string
            example: "design"
        - name: thread
          in: query
          description: Filter by thread ID
          schema:
            type: string
            pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
            example: "th-01K5WHS123EXAMPLE456789ABC"
        - name: source
          in: query
          description: Filter by message source
          schema:
            type: string
            enum: [gpts, copilot, claude, human]
        - name: author
          in: query  
          description: Filter by author
          schema:
            type: string
            example: "ChatGPT"
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /trigger:
    post:
      summary: Trigger AI agent action (Shelter Mode)
      description: |
        Triggers an AI agent action. In Shelter mode, actions are recorded 
        but not executed. Enables future automation while maintaining security.
      operationId: triggerAgent
      tags:
        - YuiFlow
        - Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTrigger'
            examples:
              summarize_request:
                summary: Summarization request
                value:
                  type: "summarize"
                  payload:
                    focus: "key_decisions"
                    format: "bullet_points"
                  reply_to: "th-01K5WHS123EXAMPLE456789ABC"
      responses:
        '200':
          description: Trigger processed (Shelter mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerResponse'
  /export/context/{thread}:
    get:
      summary: Export Context Packet for thread
      description: |
        Generates a YuiFlow-compliant Context Packet for GPTs⇄Copilot bridging.
        Aggregates fragments and knots for the specified thread.
      operationId: exportContextPacket
      tags:
        - YuiFlow
        - Export
      parameters:
        - name: thread
          in: path
          required: true
          description: Thread ID to export
          schema:
            type: string
            pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
        - name: intent
          in: query
          description: Export intent description
          schema:
            type: string
            default: "export"
      responses:
        '200':
          description: Context Packet (JSON)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextPacket'
  /export/markdown/{thread}:
    get:
      summary: Export thread as Copilot-ready Markdown
      description: |
        Exports thread conversation as human-readable Markdown optimized 
        for feeding to GitHub Copilot or other AI assistants.
      operationId: exportMarkdown
      tags:
        - YuiFlow
        - Export
      parameters:
        - name: thread
          in: path
          required: true
          description: Thread ID to export
          schema:
            type: string
            pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
        - name: metadata
          in: query
          description: Include metadata in output
          schema:
            type: boolean
            default: true
        - name: knots
          in: query
          description: Include knots (key points) in output
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Thread as Markdown
          content:
            text/markdown:
              schema:
                type: string
...
  /recent:
    get:
      summary: Get recent activity (legacy compatibility)
      operationId: getRecentActivity
      x-openai-hidden: true
      tags:
        - Legacy
      parameters:
        - name: n
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Recent items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecentItem'
components:
  schemas:
...
    InputMessage:
      type: object
      description: YuiFlow InputMessage format for /save endpoint
      required: [source, thread, author, text]
      properties:
        id:
          type: string
          pattern: '^msg-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
          description: Message ID (auto-generated if not provided)
        when:
          type: string
          format: date-time
          description: ISO8601 timestamp (defaults to now)
        source:
          type: string
          enum: [gpts, copilot, claude, human]
          description: Source of the message
        thread:
          type: string
          pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
          description: Thread ID that this message belongs to
        author:
          type: string
          description: Author of the message
          example: "ChatGPT"
        text:
          type: string
          description: Message content
        tags:
          type: array
          items:
            type: string
          description: Classification tags
          example: ["design", "feature", "decision"]
        meta:
          type: object
          description: Additional metadata
          properties:
            intent:
              type: string
              description: Intent description
            ref:
              type: string
              description: Reference to other messages
...
    ContextPacket:
      type: object
      description: YuiFlow Context Packet for GPTs⇄Copilot bridging
      properties:
        version:
          type: string
          enum: ['1.0.0']
        intent:
          type: string
          description: Intent description
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/Fragment'
        knots:
          type: array
          items:
            $ref: '#/components/schemas/Knot'
        thread:
          type: string
          pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
    AgentTrigger:
      type: object
      description: Agent trigger request
      required: [type, payload, reply_to]
      properties:
        id:
          type: string
          pattern: '^trg-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
          description: Trigger ID (auto-generated if not provided)
        when:
          type: string
          format: date-time
          description: ISO8601 timestamp (defaults to now)
        type:
          type: string
          description: Type of agent action
          example: "summarize"
        payload:
          type: object
          description: Action-specific parameters
        reply_to:
          type: string
          pattern: '^th-[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$'
          description: Thread to reply to

    # Response Schemas
    SaveResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              example: "rec-01K5WPHJ5JS0B5YYWCHETY54ZM"
            thread:
              type: string
              example: "th-01K5WPGKPT90AFEFGXQBSE23CR"
            when:
              type: string
              format: date-time
        timestamp:
          type: string
          format: date-time
    SearchResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        query:
          type: string
        filters:
          type: object
          properties:
            tag:
              type: string
            thread:
              type: string
        total:
          type: integer
        hits:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
        timestamp:
          type: string
          format: date-time
    SearchHit:
      type: object
      properties:
        id:
          type: string
        score:
          type: number
        title:
          type: string
        path:
          type: string
        snippet:
          type: string
        url:
          type: string
        date:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        thread:
          type: string
    TriggerResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        mode:
          type: string
          example: "shelter"
        ref:
          type: string
          description: Trigger reference ID
        message:
          type: string
          example: "Trigger recorded but not executed (shelter mode)"
        timestamp:
          type: string
          format: date-time
        export_url:
          type: string
          description: URL to export context for this trigger
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # Legacy Compatibility
    RecentItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        path:
          type: string
        date:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        decision:
          type: string
          enum: ["採用", "保留", "却下"]
        actors:
          type: array
          items:
            type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-yuihub-token
      description: API token for YuiHub authentication
...